Refactoring: PbfLcMsRun
Moving anything that has to do with reading from or writing to a .pbf file at the BinaryReader/BinaryWriter level to PbfLcMsRun, so that an update to the format does not require changes across multiple files and classes.
More refactoring will be made later to further break up the massive "WritePbf" function into something more manageable and comparable to the reading functions.