Don't always break mzid writing with duplicate modification names
When there are two modifications with the same name in the search, if they have the same mass/composition it is okay to ignore one when writing the mzid file - we only care about the name and the mass; if they have the same name and different masses/compositions, then throw an exception with a meaningful message.
Also make the same change in ModFileParser, and improve the exception messages there.